fix: sécurité, stabilité et performance backend (audit semaines 1-3)
Sécurité:
- JWT secret sorti du code → @Value("${jwt.secret}") / env JWT_SECRET
- DELETE /api/parametres-site et /api/tarifs-circuit-personnalise protégés ADMIN
- Mot de passe tombola → SecureRandom 24 bytes Base64
- Validation magic bytes images (JPEG, PNG, GIF, WebP) dans ImageStorageServiceImpl
- show-sql désactivé, logs Hibernate → WARN

Stabilité:
- @Valid sur @RequestBody CircuitPersonnaliseController et TarifsCircuitPersonnaliseController
- @Transactional ajouté sur 18 services write
- prixTotal et prixParNuit double → BigDecimal (entités, DTOs, services)
- Migration V6: colonnes prix en NUMERIC(10,2)
- Handlers IOException et DataAccessException dans GlobalExceptionHandler
- UtilisateurController: accès direct repository supprimé → passe par le service

Performance:
- @Cacheable sur ZoneService, VilleService, ParametresSiteService + @CacheEvict
- spring-boot-starter-cache + @EnableCaching
- Migration V7: indexes sur statuts et clés étrangères
- @EntityGraph sur DevisRepository, ReservationRepository, ReservationHebergementRepository

+++ src/main/java/com/beninexplo/backend/dto/CircuitPersonnaliseDTO.java
@@ -28,7 +28,7 @@ public class CircuitPersonnaliseDTO {
     private Long hebergementId;
     private String hebergementNom;
     private String hebergementLocalisation;
-    private Double hebergementPrixParNuit;
+    private BigDecimal hebergementPrixParNuit;
     private LocalDate dateArriveeHebergement;
     private LocalDate dateDepartHebergement;
     private Integer nombreNuitsHebergement;
@@ -169,8 +169,8 @@ public class CircuitPersonnaliseDTO {
     public String getHebergementLocalisation() { return hebergementLocalisation; }
     public void setHebergementLocalisation(String hebergementLocalisation) { this.hebergementLocalisation = hebergementLocalisation; }
 
-    public Double getHebergementPrixParNuit() { return hebergementPrixParNuit; }
-    public void setHebergementPrixParNuit(Double hebergementPrixParNuit) { this.hebergementPrixParNuit = hebergementPrixParNuit; }
+    public BigDecimal getHebergementPrixParNuit() { return hebergementPrixParNuit; }
+    public void setHebergementPrixParNuit(BigDecimal hebergementPrixParNuit) { this.hebergementPrixParNuit = hebergementPrixParNuit; }
 
     public LocalDate getDateArriveeHebergement() { return dateArriveeHebergement; }
     public void setDateArriveeHebergement(LocalDate dateArriveeHebergement) { this.dateArriveeHebergement = dateArriveeHebergement; }
